init commit fow at91sam7x with lwIP
changed port for at91sam7x512.
added ethernet driver for uIP and lwIP - new frames are not passed to
lwIP/uIP as they come, but as periodic interrupt occurs.
adder lwIP and MatrixSSL code, but not created Lua modules for them.
updated FatFS.
Contains web server which is using lwIP raw API and MatrisSSL for Lua
WSAPI pages. This server SHOULD NOT be used in anything that is just a
little bit serious, because it contains a few know bugs.